+++ spec/dummy/app/models/post.rb
@@ -2,7 +2,7 @@ class Post < ActiveRecord::Base
   stampable stamper_class_name: :person
   validates :creator, presence: true
   has_many :comments
-  
+
   # override destroy to get soft delete like acts_as_paranoid style delete
   # Note: delete_all (used in helper) bypasses this and deletes all rows.
   def destroy